+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 - `limestone/Slider` props `ticks` and `labels` to display equally spaced tick marks and labels
 - `limestone/Slider` prop `alignStepsWithTicks` to snap the knob to tick marks
 - `limestone/Slider` prop `automaticLabels` to label ticks from `min` and `max`
+- `limestone/WizardPanels` prop `noContentAnimation` to disable the content transition animation
 
 ### Changed
 
@@ -27,6 +28,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 ### Added
 
+- `limestone/IconItem` prop `description`
 - `limestone/Image` props `backgroundColor` and `backgroundSrc` to be able to change the background of a transparent image
 - `limestone/ImageItem` props `backgroundColor` and `backgroundSrc` to be able to change the background of a transparent image